ISPN-6866 OffHeap Data Container

+++ core/src/main/java/org/infinispan/marshall/core/Primitives.java
@@ -83,7 +83,7 @@ final class Primitives {
             out.writeBoolean((boolean) obj);
             break;
          case ID_BYTE_OBJ:
-            out.writeByte((int) obj);
+            out.writeByte((byte) obj);
             break;
          case ID_CHAR_OBJ:
             out.writeChar((int) obj);
@@ -101,7 +101,7 @@ final class Primitives {
             out.writeLong((long) obj);
             break;
          case ID_SHORT_OBJ:
-            out.writeShort((int) obj);
+            out.writeShort((short) obj);
             break;
          case ID_BOOLEAN_ARRAY:
             Primitives.writeBooleanArray((boolean[]) obj, out);
